+++ src/js/bundle/main.js
@@ -1,0 +1,10 @@
+'use strict';
+
+import "es5-shim";
+import "es6-shim";
+import * as ready from './ready';
+import * as language from './language';
+import * as expander from './expander';
+import * as clipboard from './clipboard';
+import * as highlight from './highlight';
+
