+++ src/js/logout/logout.js
@@ -1,4 +1,4 @@
-"use strict";
+'use strict';
 
 /**
  * This class is used for the logout page.
