+++ templates/auth_status.twig
@@ -1,10 +1,6 @@
 {% set pagetitle = 'Authentication status'|trans %}
 {% extends 'base.twig' %}
 
-{% block preload %}
-    <link rel="preload" href="{{ asset('js/main.js') }}" as="script">
-{% endblock %}
-
 {% block content %}
 <h2>{{ pagetitle }}</h2>
 
@@ -120,7 +116,3 @@
     </div>
 {% endif %}
 {% endblock %}
-{% block postload %}
-    <script src="{{ asset('js/bundle.js') }}"></script>
-{% endblock %}
-
