+++ templates/base.twig
@@ -12,7 +12,7 @@
     <link rel="stylesheet" href="{{ asset("assets/css/src/default-rtl.css") }}">
     {% endif %}
     <meta name="robots" content="noindex, nofollow">
-    <link rel="preload" href="{{ asset('js/main.js') }}" as="script">
+    <link rel="preload" href="{{ asset('js/bundle.js') }}" as="script">
     {% block preload %}{% endblock %}
   </head>
   <body id="{{ templateId }}">
@@ -30,7 +30,7 @@
     <div id="foot">
       {% block footer %}{% include "_footer.twig" %}{% endblock %}
     </div>
-    <script src="{{ asset('js/main.js') }}"></script>
+    <script src="{{ asset('js/bundle.js') }}"></script>
     {% block postload %}{% endblock %}
   </body>
 </html>
